refactor(cli): extract hooks command into commands/hooks_integration_commands.py
Move the 'hooks' command for Claude Code integration monitoring out of
the cli.py monolith into a dedicated module. cli.py now imports and
registers it via cli.add_command().

Parent-relative imports adjusted for the new module depth. All option
decorators and logic preserved as-is.

Baseline CLI test suite: 4 pre-existing failures maintained.
